refactor(report): extract report models into models/report.py
Issue #45 (EPIC-07-ST-01). Sechs Datenklassen aus report_agent.py
(ReportStatus, ReportSection, ReportOutline, Report, EvidenceItem,
ReportClaim — 131 LOC, in sich geschlossen) in eigenes Modul
backend/app/models/report.py verschoben.

- Re-Export im Service-Modul hält bestehende Caller stabil
  (api/report.py, api/runs.py, test_report_manager.py,
  test_report_export.py — keine Änderungen nötig).
- Pattern folgt models/task.py und models/project.py.
- Unused Imports dataclass + Enum aus report_agent.py raus.
- Datei schrumpft 3184 → 3053 LOC (-4 %).
- Logger und Agent verbleiben in report_agent.py (Scope #46/#47).

Verifikation: npm run check 5/5 grün — Backend 517 passed
(2 skipped Redis), Frontend 40 passed, Lint clean, Build ok.

Slice A3 von v0.9.0-Pfad A — Pfad A komplett.

Closes #45
