+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format angelehnt an [Keep a Changelog](https://keepachangelog.com/de/1.1.0/), Ve
 
 ### Geändert
 
+- **Issue #45 (EPIC-07-ST-01: Report-Models extrahieren) abgeschlossen.** Sechs Datenklassen (`ReportStatus`, `ReportSection`, `ReportOutline`, `Report`, `EvidenceItem`, `ReportClaim`) aus `backend/app/services/report_agent.py` (3184 LOC) in neue Datei `backend/app/models/report.py` (165 LOC) verschoben. Re-Export im Service-Modul hält bestehende Caller (`api/report.py`, `api/runs.py`, `test_report_manager.py`, `test_report_export.py`) unverändert. Unused Imports `dataclasses.dataclass` und `enum.Enum` aus `report_agent.py` entfernt; Datei schrumpft auf 3053 LOC (−4 %). `models/__init__.py` exportiert die neuen Klassen mit. Verhalten unverändert, 517 Backend- + 40 Frontend-Tests grün. Arbeitsprotokoll: `docu/2026-05-01-issue45-report-models-arbeitsprotokoll.md`.
 - **Issue #44 (EPIC-06-ST-04: Branching-Service extrahieren) abgeschlossen.** Neue Datei `backend/app/services/branching_service.py` (230 LOC) übernimmt `list_branches`, `create_branch` und `_apply_persona_overrides` als Funktionsmodul mit `SimulationManager` als Parameter — vermeidet zirkuläre Importe und hält die Manager-API stabil (keine Caller-Anpassungen in `api/simulation_profiles.py` oder Tests nötig). `simulation_manager.py` reduziert sich um 167 LOC auf 622 (−21 %); Manager-Methoden sind 1-Zeilen-Delegationen, unused Imports `shutil`/`Config`/`ArtifactLocator`/`RunRegistry` entfernt. `SimulationStatus`-Import in `create_branch` ist lazy, damit der Manager-→-Service-Import nicht in einen Zyklus kippt. Verhalten unverändert, 517 Backend-Tests + 40 Frontend-Tests grün. Arbeitsprotokoll: `docu/2026-05-01-issue44-branching-service-arbeitsprotokoll.md`.
 - **Issue #41 (EPIC-06-ST-01: SimulationRepository einführen) als bereits erledigt geschlossen.** Inventur zum v0.9.0-Slice-Start zeigt, dass `backend/app/services/artifact_store.py` (Issue #13, Hexagonal-Port mit `LocalFilesystemArtifactStore` und `InMemoryArtifactStore`) alle Akzeptanzkriterien retrospektiv erfüllt. `SimulationManager` (Konstruktor-DI, 7+ Store-Aufrufstellen) und `SimulationRunner` (`resolve_default_store()`) konsumieren den Store durchgängig; `state.json`-Zugriffe sind über den logischen Namen `"state"` gekapselt; File-Pfade zentralisiert in `ArtifactLocator`. Restliche `open()`-Calls im Manager betreffen Report-Meta (Folge-Issue #46) oder den Twitter-CSV-Export (out-of-scope). Status-Doku: `docu/2026-05-01-issue41-simulation-repository-status.md`.
 - **Neo4j-Memory-Settings angehoben** (`docker-compose.yml`): `server.memory.pagecache.size` 256m → **4g**, `heap.max_size` 1g → 2g, `heap.initial_size` 256m → 512m. Reaktion auf I/O-Last bei mittelgroßen Graphen — der alte 256m-Pagecache verursachte Random-Reads von der Platte. Werte sind im Compose-File die Source-of-Truth, kein `.env`-Override nötig. Rollback durch alte Werte. Live verifiziert in Neo4j 5.18 via `CALL dbms.listConfig()`.
