+++ backend/app/models/__init__.py
@@ -4,6 +4,18 @@ Data Models Module
 
 from .task import TaskManager, TaskStatus
 from .project import Project, ProjectStatus, ProjectManager
+from .report import (
+    EvidenceItem,
+    Report,
+    ReportClaim,
+    ReportOutline,
+    ReportSection,
+    ReportStatus,
+)
 
-__all__ = ['TaskManager', 'TaskStatus', 'Project', 'ProjectStatus', 'ProjectManager']
-
+__all__ = [
+    'TaskManager', 'TaskStatus',
+    'Project', 'ProjectStatus', 'ProjectManager',
+    'Report', 'ReportStatus', 'ReportSection', 'ReportOutline',
+    'EvidenceItem', 'ReportClaim',
+]
